Controllers: Convert: Fixup URL not found
Error Info :
[As the user, i want to download processed file after complete process from Extract Image only from PDF. But it was return error 404 (URL Redirect not found) instead of downloading the file]

ProcessID :
NULL

Environment :
Production

Step to reproduce :
- User open convert from PDF menu
- User upload PDF file
- User choose 'Image' with enable switch 'Extract Image Only'
- User click 'Convert PDF'

Submitter :
Andrean Alvianto

Root cause :
- In specific case iLovePDF API return static randomize filename that ignore name variable that already set it before, with that random name existed logic can’t find the correct exact file due different file name.

Resolution :
- Create new separate directory only for this specific condition to dump image with PDF only have 1 pages, after that move that completed extracted files into download directory with filename already renamed to set variable before

Related Changes :
- app\Http\Controllers\convertController.php

+++ app/Http/Controllers/convertController.php
@@ -714,14 +714,24 @@ class convertController extends Controller
                                 }
                                 $pdfUpload_Location = env('PDF_UPLOAD');
                                 $pdfProcessed_Location = env('PDF_DOWNLOAD');
+                                $pdfExtImage_Location = env('ILOVEPDF_EXT_IMG_DIR');
                                 $pdfEncKey = bin2hex(random_bytes(16));
                                 $file = $request->post('fileAlt');
                                 $pdfName = basename($file);
                                 $pdfNameWithoutExtension = basename($pdfName, ".pdf");
                                 $pdfNewPath = Storage::disk('local')->path('public/'.$pdfUpload_Location.'/'.$pdfName);
+                                $pdfTotalPages = AppHelper::instance()->count($pdfNewPath);
 						        $fileSize = filesize($pdfNewPath);
                                 $hostName = AppHelper::instance()->getUserIpAddr();
                                 $newFileSize = AppHelper::instance()->convert($fileSize, "MB");
+                                if ($pdfTotalPages = 1) {
+                                    $files = glob(Storage::disk('local')->path('public/'.$pdfExtImage_Location).'/*');
+                                    foreach($files as $file) {
+                                        if (is_file($file)){
+                                            unlink($file);
+                                        }
+                                    }
+                                }
                                 try {
                                     $ilovepdfTask = new PdfjpgTask(env('ILOVEPDF_PUBLIC_KEY'),env('ILOVEPDF_SECRET_KEY'));
                                     $ilovepdfTask->setFileEncryption($pdfEncKey);
@@ -732,7 +742,11 @@ class convertController extends Controller
                                     $ilovepdfTask->setOutputFileName($pdfNameWithoutExtension);
                                     $ilovepdfTask->setPackagedFilename($pdfNameWithoutExtension);
                                     $ilovepdfTask->execute();
-                                    $ilovepdfTask->download(Storage::disk('local')->path('public/'.$pdfProcessed_Location));
+                                    if ($pdfTotalPages = 1 && $extMode) {
+                                        $ilovepdfTask->download(Storage::disk('local')->path('public/'.$pdfExtImage_Location));
+                                    } else {
+                                        $ilovepdfTask->download(Storage::disk('local')->path('public/'.$pdfProcessed_Location));
+                                    }
                                 } catch (\Ilovepdf\Exceptions\StartException $e) {
                                     try {
                                         DB::table('pdf_convert')->insert([
@@ -974,8 +988,14 @@ class convertController extends Controller
                                         return redirect()->back()->withErrors(['error'=>'PDF Conversion failed !', 'processId'=>$uuid])->withInput();
                                     }
                                 }
-                                if (file_exists($pdfNewPath)) {
-                                    unlink($pdfNewPath);
+                                if ($pdfTotalPages = 1) {
+                                    foreach (glob(Storage::disk('local')->path('public/'.$pdfExtImage_Location).'/*.jpg') as $filename) {
+                                        rename($filename, Storage::disk('local')->path('public/'.$pdfProcessed_Location.'/'.$pdfNameWithoutExtension.'.jpg'));
+                                    }
+                                } else {
+                                    if (file_exists($pdfNewPath)) {
+                                        unlink($pdfNewPath);
+                                    }
                                 }
                                 if (file_exists(Storage::disk('local')->path('public/'.$pdfProcessed_Location.'/'.$pdfNameWithoutExtension.'.zip'))) {
                                     try {
@@ -1010,7 +1030,41 @@ class convertController extends Controller
                                         ]);
                                         return redirect()->back()->withErrors(['error'=>'PDF Conversion failed !', 'processId'=>$uuid])->withInput();
                                     }
-                                } else if (Storage::disk('local')->path('public/'.$pdfProcessed_Location.'/'.$pdfNameWithoutExtension.'-0001.jpg')) {
+                                } else if (Storage::disk('local')->path('public/'.$pdfProcessed_Location.'/'.$pdfNameWithoutExtension.'.jpg')) {
+                                    try {
+                                        DB::table('pdf_convert')->insert([
+                                            'processId' => $uuid,
+                                            'fileName' => $pdfName,
+                                            'fileSize' => $newFileSize,
+                                            'container' => $convertType,
+                                            'img_extract' => $extMode,
+                                            'result' => true,
+                                            'err_reason' => null,
+                                            'err_api_reason' => null,
+                                            'procStartAt' => AppHelper::instance()->getCurrentTimeZone()
+                                        ]);
+                                        return redirect()->back()->with([
+                                            "stats" => "scs",
+                                            "res"=>Storage::disk('local')->url('temp/'.$pdfNameWithoutExtension.'.jpg'),
+                                            'processId'=>$uuid
+                                        ])->withInput();
+                                    } catch (QueryException $ex) {
+                                        return redirect()->back()->withErrors(['error'=>'Database connection error !', 'processId'=>'null'])->withInput();
+                                    } catch (\Exception $e) {
+                                        DB::table('pdf_convert')->insert([
+                                            'processId' => $uuid,
+                                            'fileName' => 'null',
+                                            'fileSize' => 'null',
+                                            'container' => 'null',
+                                            'img_extract' => false,
+                                            'result' => false,
+                                            'err_reason' => 'Eloquent transaction error !, Catch on Exception',
+                                            'err_api_reason' => $e->getMessage(),
+                                            'procStartAt' => AppHelper::instance()->getCurrentTimeZone()
+                                        ]);
+                                        return redirect()->back()->withErrors(['error'=>'PDF Conversion failed !', 'processId'=>$uuid])->withInput();
+                                    }
+                                }  else if (Storage::disk('local')->path('public/'.$pdfProcessed_Location.'/'.$pdfNameWithoutExtension.'-0001.jpg')) {
                                     try {
                                         DB::table('pdf_convert')->insert([
                                             'processId' => $uuid,
